refactor: adjust the paragraph layout flow
