+++ packages/engine-render/src/components/docs/layout/block/paragraph/inline-drawing.ts
@@ -14,12 +14,17 @@
  * limitations under the License.
  */
 
-export * from './column';
-export * from './convert-cursor';
-export * from './line';
-export * from './liquid';
-export * from './page';
-export * from './section';
-export * from './span';
-export * from './text-range';
-export * from './tools';
+import type { IDrawing, LocaleService } from '@univerjs/core';
+
+import type { IDocumentSkeletonPage } from '../../../../../basics/i-document-skeleton-cached';
+import type { IParagraphConfig, ISectionBreakConfig } from '../../../../../basics/interfaces';
+
+export function dealWidthInlineDrawing(
+    drawing: IDrawing,
+    sectionBreakConfig: ISectionBreakConfig,
+    allPages?: IDocumentSkeletonPage[],
+    paragraphConfig?: IParagraphConfig,
+    localeService?: LocaleService
+) {
+    return [];
+}
