+++ packages/engine-render/src/components/docs/layout/block/section.ts
@@ -16,10 +16,10 @@
 
 import { DataStreamTreeNodeType } from '@univerjs/core';
 
-import type { IDocumentSkeletonPage, ISkeletonResourceReference } from '../../../basics/i-document-skeleton-cached';
-import type { ISectionBreakConfig } from '../../../basics/interfaces';
-import type { DataStreamTreeNode } from '../view-model/data-stream-tree-node';
-import type { DocumentViewModel } from '../view-model/document-view-model';
+import type { IDocumentSkeletonPage, ISkeletonResourceReference } from '../../../../basics/i-document-skeleton-cached';
+import type { ISectionBreakConfig } from '../../../../basics/interfaces';
+import type { DataStreamTreeNode } from '../../view-model/data-stream-tree-node';
+import type { DocumentViewModel } from '../../view-model/document-view-model';
 import { dealWithBlockError } from './block-error';
 
 import { dealWidthParagraph } from './paragraph';
