+++ packages/engine-render/src/components/docs/layout/model/column.ts
@@ -17,7 +17,7 @@
 import type { ISectionColumnProperties } from '@univerjs/core';
 import { ColumnSeparatorType } from '@univerjs/core';
 
-import type { IDocumentSkeletonColumn } from '../../../basics/i-document-skeleton-cached';
+import type { IDocumentSkeletonColumn } from '../../../../basics/i-document-skeleton-cached';
 
 export function createSkeletonColumn(
     columnIndex: number = 0,
