+++ packages/engine-render/src/components/docs/liquid.ts
@@ -21,8 +21,8 @@ import type {
     IDocumentSkeletonPage,
     IDocumentSkeletonSection,
     IDocumentSkeletonSpan,
-} from '../../../basics/i-document-skeleton-cached';
-import { PageLayoutType } from '../../../basics/i-document-skeleton-cached';
+} from '../../basics/i-document-skeleton-cached';
+import { PageLayoutType } from '../../basics/i-document-skeleton-cached';
 
 export class Liquid {
     private _translateX: number = 0;
