+++ packages/engine-render/src/components/index.ts
@@ -15,7 +15,6 @@
  */
 
 export * from './component';
-export * from './docs';
 export * from './extension';
 export * from './sheets';
 export * from './skeleton';
